requirements.txt
Flask
mysql-connector-python
python-dotenv
requests
